+++ src/app/routes/transactions/placeOrder.ts
@@ -110,11 +110,8 @@ placeOrderTransactionsRouter.post(
 
     ],
     validator,
-    // tslint:disable-next-line:max-func-body-length
     async (req, res, next) => {
         try {
-            // const now = new Date();
-
             // WAITER有効設定であれば許可証をセット
             let passport: cinerino.factory.transaction.placeOrder.IPassportBeforeStart | undefined;
             if (!WAITER_DISABLED) {
@@ -131,7 +128,6 @@ placeOrderTransactionsRouter.post(
                 };
             }
 
-            // const ownershipInfoRepo = new cinerino.repository.OwnershipInfo(mongoose.connection);
             const projectRepo = new cinerino.repository.Project(mongoose.connection);
             const transactionRepo = new cinerino.repository.Transaction(mongoose.connection);
 
@@ -143,57 +139,11 @@ placeOrderTransactionsRouter.post(
             });
             const seller = await sellerService.findById({ id: <string>req.body.seller.id });
 
-            const passportValidator: cinerino.service.transaction.placeOrderInProgress.IPassportValidator =
-                (params) => {
-                    // 許可証発行者確認
-                    const validIssuer = params.passport.iss === process.env.WAITER_PASSPORT_ISSUER;
-
-                    // スコープのフォーマットは、Transaction:PlaceOrder:${sellerId}
-                    const newExplodedScopeStrings = params.passport.scope.split(':');
-                    const newValidScope = (
-                        newExplodedScopeStrings[0] === 'Transaction' && // スコープ接頭辞確認
-                        newExplodedScopeStrings[1] === cinerino.factory.transactionType.PlaceOrder && // スコープ接頭辞確認
-                        (
-                            // tslint:disable-next-line:no-magic-numbers
-                            newExplodedScopeStrings[2] === req.body.seller.id || newExplodedScopeStrings[2] === '*' // 販売者ID確認
-                        )
-                    );
-
-                    // スコープのフォーマットは、placeOrderTransaction.${sellerIdentifier}
-                    // cinemasunshine対応
-                    const oldExplodedScopeStrings = params.passport.scope.split('.');
-                    const oldValidScope = (
-                        oldExplodedScopeStrings[0] === 'placeOrderTransaction' && // スコープ接頭辞確認
-                        oldExplodedScopeStrings[1] === seller.identifier // 販売者識別子確認
-                    );
-
-                    // スコープスタイルは新旧どちらか一方有効であれok
-                    const validScope = newValidScope || oldValidScope;
-
-                    // クライアントの有効性
-                    let validClient = true;
-                    if (req.user.client_id !== undefined) {
-                        if (Array.isArray(params.passport.aud) && params.passport.aud.indexOf(req.user.client_id) < 0) {
-                            validClient = false;
-                        }
-                    }
-
-                    return validIssuer && validScope && validClient;
-                };
+            const passportValidator = createPassportValidator(seller, req.user.client_id);
 
             const project = await projectRepo.findById({ id: req.project.id });
             const useTransactionClientUser =
                 project.settings !== undefined && (<any>project.settings).useTransactionClientUser === true;
-
-            // 現在所有している会員プログラムを全て検索
-            // const programMembershipOwnershipInfos =
-            //     await ownershipInfoRepo.search({
-            //         project: { id: { $eq: req.project.id } },
-            //         typeOfGood: { typeOf: cinerino.factory.chevre.programMembership.ProgramMembershipType.ProgramMembership },
-            //         ownedBy: { id: req.agent.id },
-            //         ownedFrom: now,
-            //         ownedThrough: now
-            //     });
 
             const orderName: string | undefined = (typeof req.body.object?.name === 'string') ? req.body.object?.name : DEFAULT_ORDER_NAME;
 
@@ -210,8 +160,6 @@ placeOrderTransactionsRouter.post(
                             })
                             : []
                     ]
-                    // 所有メンバーシップを記録
-                    // ...{ memberOfs: programMembershipOwnershipInfos.map((o) => o.typeOfGood) }
                 },
                 seller: req.body.seller,
                 object: {
@@ -219,7 +167,7 @@ placeOrderTransactionsRouter.post(
                     ...(useTransactionClientUser) ? { clientUser: req.user } : undefined,
                     ...(typeof orderName === 'string') ? { name: orderName } : undefined
                 },
-                passportValidator: passportValidator
+                passportValidator
             })({
                 project: projectRepo,
                 transaction: transactionRepo
@@ -234,6 +182,55 @@ placeOrderTransactionsRouter.post(
         }
     }
 );
+
+function createPassportValidator(
+    seller: cinerino.factory.chevre.seller.ISeller,
+    clientId?: string
+): cinerino.service.transaction.placeOrderInProgress.IPassportValidator {
+    return (params) => {
+        // 許可証発行者確認
+        const validIssuer = params.passport.iss === process.env.WAITER_PASSPORT_ISSUER;
+
+        // スコープのフォーマットは、Transaction:PlaceOrder:${sellerId}
+        const newExplodedScopeStrings = params.passport.scope.split(':');
+        const newValidScope = (
+            newExplodedScopeStrings[0] === 'Transaction' && // スコープ接頭辞確認
+            newExplodedScopeStrings[1] === cinerino.factory.transactionType.PlaceOrder && // スコープ接頭辞確認
+            (
+                // tslint:disable-next-line:no-magic-numbers
+                newExplodedScopeStrings[2] === seller.id || newExplodedScopeStrings[2] === '*' // 販売者ID確認
+            )
+        );
+
+        // 追加特性に登録されたwaiterScopeでもok
+        const validScopesByAdditionalProperty = seller.additionalProperty?.filter((p) => p.name === 'waiterScope')
+            .map((p) => p.value);
+        const isValidByAdditionalProperty = (Array.isArray(validScopesByAdditionalProperty))
+            ? validScopesByAdditionalProperty?.includes(params.passport.scope)
+            : false;
+
+        // スコープのフォーマットは、placeOrderTransaction.${sellerIdentifier}
+        // cinemasunshine対応
+        const oldExplodedScopeStrings = params.passport.scope.split('.');
+        const oldValidScope = (
+            oldExplodedScopeStrings[0] === 'placeOrderTransaction' && // スコープ接頭辞確認
+            oldExplodedScopeStrings[1] === seller.identifier // 販売者識別子確認
+        );
+
+        // スコープスタイルは新旧どちらか一方有効であれok
+        const validScope = newValidScope || isValidByAdditionalProperty || oldValidScope;
+
+        // クライアントの有効性
+        let validClient = true;
+        if (typeof clientId === 'string') {
+            if (Array.isArray(params.passport.aud) && params.passport.aud.indexOf(clientId) < 0) {
+                validClient = false;
+            }
+        }
+
+        return validIssuer && validScope && validClient;
+    };
+}
 
 /**
  * 購入者情報を変更する