+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - 販売者検索においてlocation.branchCodesをadditionalPropertyに自動変換するように調整
 - 販売者検索においてlocation属性を自動保管するように調整(ssktsへの互換性維持対応として)
+- 許可証の有効条件に販売者の追加特性を採用
 
 ### Deprecated